Migrate from ElasticSearch to OpenSearch
Migrate from ElasticSearch/OpenDistro and Kibana to OpenSearch and
OpenSearch Dashboards. This was done to avoid issues with
ElasticSearch license and to avoid using OpenDistro as a
replacement since it has always been intended to be a temporary
replacement.

Also, scale down the a8s logging component from three Pods with
distinct roles (master, client, data) to a single Pod with all the
three roles. This was done because it's simpler and for small
clusters we don't need to have one separate Pod for each role (the
extra overhead is not worth it).

Also, move the FluentD and Fluentbit config from a ConfigMap to a
K8s Secret since that config can contain secrets such as passwords.