Fix schedule of automated image update
A previous PR merged a commit to automatically update container
images every night. That commit made the automatic update run
at a time convenient for testing, i.e., shortly after the time
at which we merged the PR, so that we could verify that the update
was successful. This commit restores the schedule to its normal
value (late night).