Use a unique namespace for the testing namespace
Our test suites delete the testing namespace once they complete.
When running all test suites at once with a testing namespace set the
the second test suite will fail due to the testing namespace being in a
terminating state.

A quick fix is to add a unique suffix to testing namespace so that this
issue does not occur. This may not be what we want since it changes
expectations the user or automation (LATR) may have about the testing
namespace they believe to be targeting.

Some other approaches that come to mind:
1. Have the user or LATR delete the namespace after use. And make it so
   the testing framework does not do this. This takes responsibility away
   from individual test suites to delete the namespace after use.
2. Provide a switch via environmental variable to enable or disable
   deleting namespace at the end of the suite. This in a sense gives us
   both options.
3. Simply create a unique namespace by adding a random suffix to the
   provided testing namespace name. That is what this PR currently does.
4. Make it optional whether we add a random suffix via environtal
   variable.
5. Dynamically Use a random suffix only if the namespace is in a
   terminating state.