Add OLM resources and instructions
Resources for deployment with the OLM are created with documentation on
how to use them.

This required a restructuring of the /deploy/a8s directory, all
backupstorage related configuration is now under /backup-config, the
other files of the static manifests reside in /manifests and everything
OLM related is located under /olm.

Additionally the ConfigMap backup-store-config is renamed to
a8s-backup-store-config to be more consistent with the naming of the
secret.