Update opensearch, fluentd and fluentbit
Opensearch and opensearch dashboards have been updated to v2.0.0, this
required to update the dashboards hostname to a valid hostname in the
configuration file.

Additionally fluentd had to be updated which also introduces the
opensearch plugin, otherwise the logs cannot be forwarded due to a
deprecated field. This also required to update the fluentd configuration
to use the opensearch plugin.

To ensure compatibility fluentbit is updated too.