Failure case tests for TaintWorkers & UntaintAll
Tests for failure cases are added for the functions TaintWorkers
and UntaintAll, testing correct behavior in potential failure
scenarios, such as testing the behavior if a node update or list
of the nodes fails.

Additionally, new custom error types are introduced that contain
the affected node or node name. Test cases for ensuring
correctness of the error messages are also added.