Provide stable node to Availability Zone mapping
Previously when trying to label nodes with AZs, upon retry a
different ordering in the NodeList could cause the nodes to
be distributed not evenly among AZs, leading to having 0
nodes in one AZ while more nodes in another one.

This commit fixes that by ensuring the mapping from nodes to AZs
is stable across retries and nodes are spread evenly between AZs.