Bump PostgreSQL Operator
Bump PostgreSQL Operator to commit
8dc82ca8e22b24ac8ab424e40fae69c1f259c6ef .

Update the changelog, the PostgreSQL CRD, the generated API docs
and the go.mod of the tests accordingly.

The only user visible change introduced by this commit is that the
description of some fields that affect affinity and anti-affinity
for the Postgresql pods no longer describes said fields as beta,
thus presenting them as stable to the user.